style(msvc): fix C4244/C4267 in every remaining 3-site file (item 2)
All static_cast, documenting existing intentional narrowing -- no
logic changes. sm_tests unchanged at 5966/226; [NotesLoader]-tagged
tests (587/34) re-checked identical.

19 files, 57 sites: WheelNotifyIcon.cpp, UnlockManager.cpp, Steps.cpp
(SS5-protected, characterization-only Lua binding), ScreenSelectCharacter.cpp,
ScreenSelect.cpp, Screen.cpp, RageThreads.cpp,
RageSoundReader_ChannelSplit.cpp, RageFileDriverReadAhead.cpp,
RageFileDriverDirect.cpp, PlayerStageStats.cpp, OptionRowHandler.cpp,
NotesLoaderSSC.cpp (SS5-protected, characterization-only), LightsManager.cpp,
InputQueue.cpp, GameManager.cpp, ArrowEffects.cpp, ActorFrame.cpp,
ScreenOptionsEditCourse.cpp (a genuinely separate file from
ScreenOptions.cpp/Course.cpp, never touched before).

Every site is a size_t/int64_t/double/lua_Integer-family value
narrowed into a smaller int/unsigned/float/lua_Number type at a
wrap() call, a reverse-loop bound, an iterator-diff, or a Lua
push/rawset binding -- same patterns as the whole sweep.

+++ src/GameManager.cpp
@@ -3432,7 +3432,7 @@ int GameManager::GetIndexFromGame( const Game* pGame )
 	for( std::size_t g=0; g<ARRAYLEN(g_Games); ++g )
 	{
 		if( g_Games[g] == pGame )
-			return g;
+			return static_cast<int>(g);
 	}
 	FAIL_M(ssprintf("Game not found: %s", pGame->m_szName));
 }
@@ -3549,11 +3549,11 @@ public:
 	{
 		std::vector<const Game*> aGames;
 		p->GetEnabledGames( aGames );
-		lua_createtable(L, aGames.size(), 0);
+		lua_createtable(L, static_cast<int>(aGames.size()), 0);
 		for(std::size_t i= 0; i < aGames.size(); ++i)
 		{
 			lua_pushstring(L, aGames[i]->m_szName);
-			lua_rawseti(L, -2, i+1);
+			lua_rawseti(L, -2, static_cast<int>(i+1));
 		}
 		return 1;
 	}
