+++ src/RageFileDriverReadAhead.cpp
@@ -108,7 +108,7 @@ void RageFileDriverReadAhead::FillBuffer( int iBytes )
 	/* Seek back to where we were.  If we're going back to the cached region, seek past it,
 	 * like SeekInternal does. */
 	if( iOldPos < (int) m_sBuffer.size() )
-		iOldPos = m_sBuffer.size();
+		iOldPos = static_cast<int>(m_sBuffer.size());
 	m_pFile->Seek( iOldPos );
 
 	/* Now that we're done moving the file pointer around, set the file's read-ahead hint,
@@ -122,7 +122,7 @@ int RageFileDriverReadAhead::ReadInternal( void *pBuffer, std::size_t iBytes )
 	if( m_bReadAheadNeeded && m_iFilePos < (int) m_sBuffer.size() )
 	{
 		// If we can serve data out of the buffer, use it.
-		iRet = std::min( iBytes, m_sBuffer.size() - m_iFilePos );
+		iRet = static_cast<int>(std::min( iBytes, m_sBuffer.size() - m_iFilePos ));
 		memcpy( pBuffer, m_sBuffer.data() + m_iFilePos, iRet );
 	}
 	else
@@ -162,7 +162,7 @@ int RageFileDriverReadAhead::SeekInternal( int iOffset )
 		/* This assumes that seeking the file won't block.  This seems to be true in Linux, at least.
 		 * Seek the actual file to just past our buffer, so the RageFileManagerReadAhead::ReadAhead
 		 * call will read ahead from the correct position. */
-		m_pFile->Seek( m_sBuffer.size() );
+		m_pFile->Seek( static_cast<int>(m_sBuffer.size()) );
 		m_iFilePos = iOffset;
 		return iOffset;
 	}
