+++ src/ScreenSelect.cpp
@@ -38,7 +38,7 @@ void ScreenSelect::Init()
 	RString choice_names= CHOICE_NAMES;
 	if(choice_names.Left(4) == "lua,")
 	{
-		RString command= choice_names.Right(choice_names.size()-4);
+		RString command= choice_names.Right(static_cast<int>(choice_names.size())-4);
 		Lua* L= LUA->Get();
 		if(LuaHelpers::RunExpression(L, command, m_sName + "::ChoiceNames"))
 		{
@@ -51,7 +51,7 @@ void ScreenSelect::Init()
 				std::size_t len= lua_objlen(L, 1);
 				for(std::size_t i= 1; i <= len; ++i)
 				{
-					lua_rawgeti(L, 1, i);
+					lua_rawgeti(L, 1, static_cast<int>(i));
 					if(!lua_isstring(L, -1))
 					{
 						LuaHelpers::ReportScriptErrorFmt(m_sName + "::ChoiceNames element %zu is not a string.", i);
@@ -63,7 +63,7 @@ void ScreenSelect::Init()
 						mc.ApplyCommitsScreens(false);
 						mc.m_sName = ssprintf("%zu", i);
 						Commands cmd= ParseCommands(com);
-						mc.Load(i, cmd);
+						mc.Load(static_cast<int>(i), cmd);
 						m_aGameCommands.push_back(mc);
 					}
 					lua_pop(L, 1);
